Fixed time argument type is segment drawer. A few minor changes.

+++ LSystem/LSystem/LSystem.h
@@ -6,7 +6,7 @@
 
 @protocol SegmentDrawer <NSObject>
 
--(void) segmentFrom:(CGPoint)from to:(CGPoint)to generation:(NSInteger)generation time:(NSInteger)time identifier:(NSInteger)identifier;
+-(void) segmentFrom:(CGPoint)from to:(CGPoint)to generation:(NSInteger)generation time:(CGFloat)time identifier:(NSInteger)identifier;
 
 @end
